Standardize between the skill and misc. ability data models
