+++ templates/item/parts/item-attributes.hbs
@@ -22,7 +22,7 @@
 
 	<!-- Grid -->
 	{{#if attributePartials.grid}}
-		<section class="grid grid-2col gap-5">
+		<section class="fu-form__property-groups">
 			{{#each attributePartials.grid as |part|}}
 				{{> (part) @root}}
 			{{/each}}
